Criação dos serviços

+++ src/model/dao/AlunoDao.java
@@ -6,5 +6,5 @@ public interface AlunoDao {
 	void insert(Aluno obj);
 	void deleteById(Integer id);
 	Aluno findById(Integer id);
-	Aluno findByMatricula(Integer matricula);
+	Aluno findByMatricula(long matricula);
 }
